+++ web/netlify/functions/add-liquidity-background.ts
@@ -8,9 +8,9 @@ import { PrivateKeyAccount, erc20Abi, zeroAddress } from "viem";
 import { Address, privateKeyToAccount } from "viem/accounts";
 import { LiquidityManagerAbi } from "./utils/abis/LiquidityManagerAbi";
 import { SDaiAdapterAbi } from "./utils/abis/SDaiAdapterAbi";
-import { waitForContractWrite } from "./utils/common";
 import { S_DAI_ADAPTER, liquidityManagerAddressMapping } from "./utils/common";
 import { config } from "./utils/config";
+import { waitForContractWrite } from "./utils/waitForContractWrite";
 
 export default async (req: Request) => {
   const [chainIdString, marketId] = req.url.replace(/\/$/, "").split("/").slice(-2);
