+++ web/netlify/functions/scheduled-notification-events-sender.mts
@@ -1,9 +1,9 @@
 import { createClient } from "@supabase/supabase-js";
-import { getPostmarkClient } from "./utils/common";
+import * as postmark from "postmark";
 import { FROM_EMAIL } from "./utils/common";
 
 const supabase = createClient(process.env.VITE_SUPABASE_PROJECT_URL!, process.env.VITE_SUPABASE_API_KEY!);
-const postmarkClient = getPostmarkClient();
+const postmarkClient = new postmark.ServerClient(process.env.POSTMARK_API_TOKEN!);
 
 async function sendPendingNotifications() {
   try {
