+++ web/package.json
@@ -103,7 +103,7 @@
     "vite": "^6.0.0",
     "vite-plugin-cjs-interop": "^2.1.4",
     "vite-tsconfig-paths": "^5.1.4",
-    "vitest": "^3.1.1"
+    "vitest": "3.1.4"
   },
   "workspaces": {
     "nohoist": [
