+++ web/src/components/Market/Header/QuestionLine.tsx
@@ -2,13 +2,13 @@ import { useArbitrationRequest } from "@/hooks/useArbitrationRequest";
 import { useReopenQuestion } from "@/hooks/useReopenQuestion";
 import { useResolveMarket } from "@/hooks/useResolveMarket";
 import { SupportedChain, mainnet } from "@/lib/chains";
+import { getTimeLeft } from "@/lib/date.ts";
 import { CheckCircleIcon, HourGlassIcon, RightArrow } from "@/lib/icons";
 import { MarketTypes } from "@/lib/market";
 import { MarketStatus } from "@/lib/market.ts";
 import { Market, Question } from "@/lib/market.ts";
 import { paths } from "@/lib/paths";
 import { ANSWERED_TOO_SOON, getAnswerText, getQuestionStatus, getRealityLink, isFinalized } from "@/lib/reality";
-import { getTimeLeft } from "@/lib/utils";
 import clsx from "clsx";
 import { COLORS } from "./index.ts";
