+++ web/src/lib/utils.ts
@@ -1,8 +1,3 @@
-import { intervalToDuration } from "date-fns";
-import { formatInTimeZone } from "date-fns-tz";
-import { compareAsc } from "date-fns/compareAsc";
-import { FormatDurationOptions, formatDuration } from "date-fns/formatDuration";
-import { fromUnixTime } from "date-fns/fromUnixTime";
 import { formatUnits, getAddress } from "viem";
 import SEER_ENV from "./env";
 
@@ -13,51 +8,6 @@ export const INVALID_RESULT_OUTCOME = "Invalid result";
 
 // display text for invalid outcome
 export const INVALID_RESULT_OUTCOME_TEXT = "Invalid";
-
-// https://stackoverflow.com/a/72190364
-export function localTimeToUtc(utcTime: Date | string | number) {
-  if (typeof utcTime === "string" || typeof utcTime === "number") {
-    // biome-ignore lint/style/noParameterAssign:
-    utcTime = new Date(utcTime);
-  }
-
-  const tzOffset = utcTime.getTimezoneOffset() * 60000;
-  return new Date(utcTime.getTime() - tzOffset);
-}
-
-export function formatDate(timestamp: number, formatString?: string) {
-  const date = fromUnixTime(timestamp);
-  return formatInTimeZone(date, "UTC", formatString ?? "MMMM d yyyy, HH:mm");
-}
-
-export function getTimeLeft(endDate: Date | string | number, withSeconds = false): string | false {
-  const startDate = new Date();
-
-  if (typeof endDate === "number" || typeof endDate === "string") {
-    // biome-ignore lint/style/noParameterAssign:
-    endDate = fromUnixTime(Number(endDate));
-  }
-
-  if (compareAsc(startDate, endDate) === 1) {
-    return false;
-  }
-
-  const duration = intervalToDuration({ start: startDate, end: endDate });
-
-  const format: FormatDurationOptions["format"] = ["years", "months", "weeks", "days", "hours"];
-
-  if (withSeconds) {
-    format.push("minutes", "seconds");
-  } else if (Number(duration.days || 0) < 1) {
-    if (Number(duration.minutes || 0) < 2) {
-      format.push("seconds");
-    } else {
-      format.push("minutes");
-    }
-  }
-
-  return formatDuration(duration, { format });
-}
 
 export function shortenAddress(address: string): string {
   try {
